Adding support for choosing curve for prod EEK
In addition to curve 25519, curve P256 is also a valid choice for
signing and ECDH within the RKP scheme. This change modifies the API for
fetching the EEK based on which curve is requested. This is primarily
motivated to support the factory extraction tool, since it currently
only provides the curve 25519 EEK certification chain.

This change does not fully implement all underlying functionality
related to the potential for using Curve P256 in this scheme. A separate
change adding validation of that curve to VTS will be coming shortly.

Bug: 215445120
Test: atest VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: Ie8ad1f3d4e56864ad2a28f93d9ec32d78b293474