Include the CWT validation in VTS testing.
This change modifies the VTS tests to validate the structure of the CWT
which contains the public key as one of several fields in the map. It
only validates the subset of fields expected to be seen in the VTS
tests. The ones related to the actual BCC will not be present in test
mode due to the security guarantees of the interface.

Additionally, the test will also now check that the first key in the BCC
array matches the public key that should be self-signed in the
certificate located in the second index of the BCC array.

Test: atest VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: I9e7769861529b8095d77ce8fee3c78222579d38c