Update UsageSetting Cache before comparison in VTS
The final step of the RadioNewtorkTest#setUsageSetting
is to restore the original value to the modem in case it was
changed. After doing that, the locally cached value in the
VTS layer needs to be updated before comparing it with the
originally queried value to ensure that the modem was
restored to its initial state.

Bug: 211819471
Test: m VtsHalRadioTargetTest
Change-Id: Ibdd584d91b16bb779a0cefa2f45d2704cdcf7288