Move vts/OWNERS files to vts/functional/OWNERS
Move these files only if vts/functional already exists.
When that functional directory exists, the VTS test module is defined
within. The OWNERS file is for that test module, so it needs to be in
the same directory that the test module is.

Adjust all of the relative paths of the moved OWNERS files.
`../../` -> `../../../`

Adjust included path to point to new OWNERS file.

Fix wifi/suplicant vts OWNERS files, they were pointing to a
non-existant file.

Remove emails that no longer exists.

This was causing issues with Android Test Ownership Service.

Bug: 222305378
Test: N/A
Change-Id: Ifac6954a1240943652d15ac40277e361e32ae455