Updated the description for APPLICATION_ID and APPLICATION_DATA
As the signature of the getKeyCharacteristics() does not
use Tag Mechanism for app_id and app_data, there is no way
to distinguish between appId / appData  values that are
absent, vs values that are present but of zero length. So
to overcome this confusion generateKey(), importKey(),
importWrappedKey(), begin() and upgradeKey() shall return ErrorCode::INVALID_TAG if appId / appData tags are present
in KeyParamters, but with zero length.

Test: VtsAidlKeyMintTarget
Change-Id: I145dcba878171c174d48ad42fadeb49e045b5c55