Allow TAG_OS_PATCHLEVEL to be less than or equal to os_patch_level
In general, TAG_OS_PATCHLEVEL should be equal to os_patch_level()
reported from the system.img in use. But it is allowed to boot a
GSI system.img with newer patch level, which means TAG_OS_PATCHLEVEL
might be less than or equal to os_patch_level() in this case.

Signed-off-by: Rohit Kothari <r.kothari@samsung.com>
Change-Id: Id10443e0fb62329db778d34db7574e330b6a7bc6