Don't mark SAT/DT as failures
If SAT/DT is not supported, should be marked as
ignored but in our case they are marked as failures.
If no support for SAT/DT the implementation returns INVALID_ARGUMENTS
and test is calling the printSkipped function.
Instead of calling printSkipped, if GTEST_SKIP() called
unsupported TC results are parsed properly and test cases will move
to ignored category

Bug: 191229970
Test: VtsHalBroadcastradioV1_0TargetTest

Signed-off-by: V,Anilkumar <anilkumar.v@intel.com>
Change-Id: I4ed7828757a6cc8b83bae989e11289da1a3e814f