Revert "audio: use binder threadpool"
This reverts commit d90c47420af3aa59c6fc864c30023ec62b0a26ac.

This change is not needed and contains an invalid behavior.
Calls to `ABinderProcess_...` are NDK aliases for methods
of `::android::ProcessState` class. Thus, this change attempts
to reconfigure and start the thread pool after it has already been
started by the code in the preceding lines.

Note that the call `ABinderProcess_setThreadPoolMaxThreadCount(1)`
decreases the thread pool capacity from the default 15 threads
to 1. In fact, ProcessState::setThreadPoolMaxThreadCount has a
check to prevent that, however it does not fire because it accesses
shared variables without taking a lock.

Change-Id: I8caaededad15f3f3f8281041f5906bc618a26c56