Clarify error message in DeviceInfo check.
This change gives additional context to the device info checks while
also stopping test execution early. Currently, an unprovisioned device
will generate a massive spam of failures which may be WAI for an early
hardware revision device that was not provisioned with attestation IDs.

This change halts the test with a failure upon the first missing entry
in the map and gives additional guidance around what may or may not be
going wrong.

Test: atest VtsHalRemotelyProvisionedComponentTest
Change-Id: I16069dba841a90aa55781148d3c268ced635e006