Make IRemotelyProvisionedComponent aware of AVF
In order to support VM attestation by the Android Virtualization
Framework (AVF), add a new "avf" certificate type. This introduces the
need to restrict which components in a system are allowed to reqeust
different certificate types. Those permissions are expressed with a new
config field in the DICE certificates.

The permissions are needed, at a minimum, to prevent an untrusted VM
requesting "avf" certificates that should only be available to a trusted
VM.  It also seems prudent to be able to prevent less privileged VMs
from requesting typically more privileged "keymint" certificates as an
extra safety precaution.

Bug: 241429277
Test: n/a comments only
Change-Id: Ia55dd9d3a09bcba5f81bd7c1d9532ec33e27d5c6