Add upgradeKey to IRemotelyProvisionedComponent
This change adds the upgradeKey method to the IRPC HAL interface. A
generic system component, RKPD, is being created to take over the
responsibilities of managing IRPC interfaces and the keys generated from
them. This component will need a generic upgradeKey method to call on
the corresponding IRPC HAL instances it manages, otherwise it will need
to have context specific code for every interface that may call into it.

Bug: 243069973
Test: The documentation is intelligible
Change-Id: Id0dc0a956d9870387032799b09d0c4b5663935ae