Revert "Remove OMX from compat matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f45f6d54-aa99-4e61-958c-b04304163a20). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic59bded63e130a54e8f0a5913584837dc3fdf759