Revert "Remove OMX from compat matrix"
Revert "Remove OMX from compat matrix"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/b45611bb-b807-4598-883a-96b226cbe40d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07895c7c24e0c5758f1606c9e47e975fbbd470ba