Revert "Add TEST_MAPPING"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2f182d13-5d23-414a-bca3-7f148eed3adf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46335e59048dd3eec4bc1765bc577f672ae59425