Revert "Add TEST_MAPPING"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/d7eb0dd3-55f8-418b-84a3-f61cdccf69fa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b9349f9614691b31135f67e7bc98feeeaccc566