Revert "Add TEST_MAPPING"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/d209389b-9f20-469a-acc4-f2f76dc057ff). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ea460752e82df267e8fc06880a82d232978c7dd