Revert "Remove OMX from compat matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/e9374d58-712a-413d-bbc0-94e148c353ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ebb4f61036b794ea019679513fc7ba566e4b432