Revert "Update VtsHalRemotelyProvisionedComponentTargetTest to v3"
Revert "keystore2_test: Update to IRPC v3"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/163e465c-4481-4c06-9f2f-cbdfb84ec682). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I833976678e2df26ad360b51c17f16bbe21d0af2f