Revert "Bump Identity and KM AIDL API versions"
Revert "keystore2_test: Update to IRPC v3"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/163e465c-4481-4c06-9f2f-cbdfb84ec682). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie9fdbda50d21524d5ae55275befafd3812c64413