Revert "Update VtsHalRemotelyProvisionedComponentTargetTest to v3"
Revert "keystore2_test: Update to IRPC v3"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/9d040e30-b1bf-4ff5-a86a-9c28d25b59d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iefb72f32038f10af3e1bef197f4046de13bbc4d6