Revert "Bump Identity and KM AIDL API versions"
Revert "keystore2_test: Update to IRPC v3"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/6b6565e4-1ef6-4bbc-adc5-cd65b688bc76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2960527e7e587181017681bf23cac3aad3a2d0c8