Revert "Bump Identity and KM AIDL API versions"
Revert "keystore2_test: Update to IRPC v3"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/7aac4370-d7de-4279-b243-5eb6dfe42b12). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icfbc1a7250947a16e579b9c9e602287dfbbb85f8