Revert "Update VtsHalRemotelyProvisionedComponentTargetTest to v3"
Revert "keystore2_test: Update to IRPC v3"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/b25ad31b-93c6-48a9-a8a1-09c74cab417e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I495fa97087e3ddf6c1abfd03829bccadf56b2fd6