Revert "Update VtsHalRemotelyProvisionedComponentTargetTest to v3"
Revert "keystore2_test: Update to IRPC v3"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/fa05f012-9965-4a20-ac27-bcbb88055bd2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I106ee93793406aac050e564adcc7f9fb8ec46cb4