Revert "Update VtsHalRemotelyProvisionedComponentTargetTest to v3"
Revert "keystore2_test: Update to IRPC v3"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/47d36b70-2961-4615-9283-b27a4937e130). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I682edf385d3b0431ab13e0275e11aa31775798e7