Revert "Remove CEC HIDL from compatibility matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/1f0441a8-c225-46e4-a372-77fb4990209c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I016807aad4d4fcc54c2be3b74fa7e25bc3e3548e