Revert "Fix tunerhal vts feId uninitialized on getFrontendIdByType"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/4dd1e214-c427-4294-9fe9-0ab0853dd0c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6980acf73693e8cf2b53c3ec5d0c3f4667fc4f19