Revert "Add AIDL files for cas and cas/native modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/b6f41c4d-e118-4818-a8a9-b0f85af6acf4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id2da9df5de63d1ebc7cdf506054f1f3c7e4ac01c