Define a limit on number of keys in CSR
Rationale here is that many IRPC implementations are memory constrained.
We define a common bound to guarantee consistent behavior across
different devices.

The specific value was chosen because 10 keys is about as much as IRPC
v2 SE implementaiton on Pixel can handle. So I'm using that as a
reference for other SEs. Another consideration here is that IRPC v3 CSR
generation request and response messages with 10 keys fit within one
page of memory.

Bug: 254137722
Test: atest VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: Ib6fb6d6ec7c74004524a5505a37aa82c9e44ef91