Add an implementation defined limit on number of keys in CSR
Rationale here is that many IRPC implementations are memory constrained.
We add a way for implementations to report the maximum number of
supported keys. This way we can guarantee consistent behavior across
different devices.

We define the lowest number of keys that implementations must support to
be 20. This specific value was chosen because the current implementation
of RemoteProvisioner already combines keys into batches of exactly 20,
i.e. this lower bound can be enforced retroactively on on older versions
of the interface.

Bug: 254137722
Test: atest VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: Ib6fb6d6ec7c74004524a5505a37aa82c9e44ef91