Raise challenge and device info into AuthenticatedMessage
The challenge is against the DICE signer and the device info adds
further information about the device that can be used for more
authentication policy decisions. Move both of these out of the
endpoint-specific message and up into the common authentication wrapper.

Simplify the versioning by having the CSR version continue sequentially,
making the current version 3. Reset the AuthenticatedMessage version
back to 1 as it isn't used to distinguish v2 and v3 requests anyway and
starting from 1 avoid confusion with the CSR version which is already
past that.

Bug: 250910137
Test: n/a -- comments only
Change-Id: I13836e90fa76b1b22cb6627f3d987828ffeb0adc