Adjust CSRv3 CDDL after implementation experience
Rename from AuthenticatedMessage to AuthenticatedRequest in order to
make the direction of the message clear.

Move the challenge out of the endpoint-specific message and up into the
common authentication wrapper as it is uesd in the authentication
protocol.

Simplify the versioning by having the CSR version continue sequentially,
making the current version 3. Reset the AuthenticatedMessage version
back to 1 as it isn't used to distinguish v2 and v3 requests anyway and
starting from 1 avoid confusion with the CSR version which is already
past that.

Bug: 250910137
Test: n/a -- comments only
Change-Id: I13836e90fa76b1b22cb6627f3d987828ffeb0adc