composer: vts: Init display dimension using active config
Dimension was initialized to the size of the last config that is in the config list.
VTS didn't change the dimension in the readback test and allocated buffer
for the readback with the dimension size.
However it was different with actual display size.
This change initializes display dimension to the actual display size.

Change-Id: I64fec02f5386eb6ca1f41ab376be193231810b93
Signed-off-by: HyunKyoung <hk310.kim@samsung.com>