Add a convert library for sensor AIDL
This is a direct copy/paste of the conversion functions from
frameworks/native/services/sensorservice/AidlSensorHalWrapper.cpp.
This conversion code needs to be shared between sensorservice, a new
libsensorserviceaidl, and the new VTS tests for AIDL sensorservice.
It will likely also be used by vendors interacting with the new AIDL
sensorservice.

Test: VtsHalSensorManagerV1_0TargetTest
Bug: 205764765
Change-Id: Iabb0bcf0770672671e51fd304f79e14ca94a2d15