Revert "IMapper 5 - the Stable C approach"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c1b1f63-6ef3-48a4-b9e9-cb7028fc6cc0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie19536b8323bf795af948a384c5cf7ef1c16a37b