Revert "IMapper 5 - the Stable C approach"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ffde8f40-05ad-47dd-aa39-6e56ca7ce7e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If783d79ba62a6d03c05a8682de8ef92c0f99988c