Revert "powerstats: add new EnergyConsumerType for camera"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/222d1597-03a6-49d2-9d6c-7339b904b4dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46d7886253f5e9b1edade2b67c736a2e72f5a44c