Revert "powerstats: add new EnergyConsumerType for camera"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6647422e-0c09-4115-ae6f-4e0b5f335042). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I01ac06071b917fdea4d79c9df1109a86ed8a3bc1