Revert "powerstats: add new EnergyConsumerType for camera"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c0206a7-940f-43dc-9c7b-736257b544cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I81ecf2e1a45fb30b32054632b85eeb767b2c04ea