Revert "powerstats: add new EnergyConsumerType for camera"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/817c421b-885e-4d0f-92a2-efbfee40faf9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4064a14a5542b8b52351e974ae2c93696757165