Revert "identity: Add support for ECDSA auth and don't require s..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9773af0a-8057-44b6-8ee9-e2ca3669ef8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7aa80c1c0848e3651f9b90cd2478c6206162e436