Revert "EnvironmentalReverb: Add AIDL placeholder implementation..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/003f15fe-d8c2-4a2b-a570-6d65c805bc3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idaa3d5518b9cad39ff14b61c64a9bc1d614dd554