Revert "EnvironmentalReverb: Add AIDL placeholder implementation..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/946c1179-5f95-47e9-a098-e710c96a02d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaadcc2788feb44e48ec88d8e1201bc1ceadce96f