Add proposed "serialno" field to DeviceInfo CBOR description
This field will be used for user feedback during device registration.
Currently when the device info server needs to address a failed
registration its best identifier is the entire CSR payload. For OEMs
that have the serial number populated by the time devices are registered
with RKP we will be able to reference devices by serial number in
addition to the CSR payload.

This will not need to be stored anywhere in Spanner. Only the user that
has uploaded the CSRs at time of registration could see the device
serial number in the server's JSON response body.

Bug: 264302050
Test: Comment change only. Only visually inspected.
Change-Id: Ib28c1e4091c08803f637b4db7e0c3c3fa8698ca1