Revert "vib example client on product"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/edf77302-b469-413d-b99e-e3962ea0301d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e6dcfb1e83da6eb48bb160ca3731ee2737fd19e