Revert "vib example client on product"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12af49c5-f071-4764-9b11-969df981c011). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I45d6d1fa0d479dd7ecc6b859f64f74fbfd55894f