Revert "vib example client on product"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/137e1475-f9f8-4748-8d47-e1d3b705a0bd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2ab00cac611d8e2e554210edc285a34e3c8f3f69