Revert "audio: clean up OWNERS"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/000004b0-e80a-4aa8-a873-531310f1f210). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib28068c073c7d8e9537fe22340607622125e1127