cameraservice: Fix prefix length path for External Camera HAL
The defauly External Camera HAL accidentally used `sizeof` on a `char *`
instead of `char[]` to get the length of the string. This "worked"
because the string in question "/dev/" has length of 5 while on a 32-bit
system, `char *` has a size of 5. This offset of 1 was handled a few
lines later.

However on 4-bit system this method no longer works and `sizeof(char *)`
returns 8. This breaks the logic later in code.

As `sizeof` was being used as a proxy for strlen, this CL converts the
`char *` into a `char[]`. Using `sizeof(char[])` returns the number of
bytes occupied by the array, which is one more than the length of the
string (for the null character).

Bug: 265168485
Test: Manually tested that the calculated length is correct after the
      fix.
Change-Id: I92481a2bcb74fdbbb483ef552168debf07b2d68a