Revert "audio VTS: Limit mix ports that route to attached devices"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/411b6a35-6164-430a-859a-b4471ed9f665). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I301de83d4eda2bb2845ddb4b7361e7776264f0f4