Revert "bluetooth: Add support for linux bluetooth raw hci sockets"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b68dd3fa-1238-4b18-b603-65cf134ca943). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c010d27168a8893719b34f4aa3baf95fa89e8fc