Revert "IMapper 5 - the Stable C approach"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd8b0b85-865c-4354-8a6d-ca722490f9e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id87159b854eac8995f7cf03f2fd2013ebb4fe9b2