Change optional="false" to "true" in compat matrix 7
A common workflow for vendors is to develop their new launching devices
targeting N-1 compatibility matrix so they can update each of their HALs
independently at different times.

This works because the target (N-1) framework compatibility matrix is squashed
with the newer compatibility matrices allowing the use of newer
interfaces.

This doesn't work when converting a service from HIDL to AIDL when the
HIDL service is marked as optional="false". It allows for the use of the
new AIDL interface, but it still requires the old HIDL interface.

Marking the old HIDL interfaces as optional allows vendors to develop
before the device targets level 8.

Test: m
Bug: 247575800
Change-Id: I4446cdaa4a44b4dae1e9eee68089bf723dd25276