Revert "Audio AIDL : Add placeholder implementation and VTS test..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78fde1c4-5139-4d1c-b6fa-23672d0d0099). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa9c3f511b4cabfca00281bdf6860313d3020f6d