Revert "Audio AIDL : Add placeholder implementation and VTS test..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50e4e6ac-6d81-4716-aaee-3c02ce6736ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21c32162fd3eb2f25e4577172f3dad99c6f08c7e