Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b5c7eb3-eeab-49c9-838b-f48b6288d66a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a61f9057201c2582f2332cb7f41bec840d4aa2f