Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b54dbdd-88e9-494d-aaf8-1cd586ae8317). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ce4b47053fb6226129d55a6924124e001d5d94f