Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/40bbdbc7-441e-4fb7-8f7e-5555a7c503d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I398f324db0bf3f605d2df489e2f98125b1e6a1df