Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f4d4470-02d0-4525-bbd7-435cb6d17f29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic0f95c7a6cc877eebf6057eff870bc9172d7d065