Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3fdc88db-15dc-470f-a3b7-93048a77f859). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ebd81bc4b9a298edd5b306a57269973b2d6c0cf