Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82e5cf81-6085-4762-abf7-2abad741ba83). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If9bc82371b21a4481893d747ae4b4178d6c34b3d