Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7d0a1cbb-604f-431d-8cb8-b64dd3e44e1f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39e6d62244c49ae26742fb0924e051f74badf49e