Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8d27bf36-7899-4fac-a6d8-5314eaccbf00). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0a41d86eb06069f7155ce33ef6af5789e704d03