Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f3fafd6-f756-4b17-9bd4-0af0eec27889). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3f0f3fdb56f338a153d78aedfc63c2068ee4049c