Revert "Add 6.1 kernel to current matrix"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a171ee9-daf3-4521-87e6-0fa8ab7feb01). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida9bb31c4e127d4d1aed449b1438862a0e2674f0