rkp: add keymint v2 hash into rkp v2
When we split the rkp interface, we only added one hash to rkp v2, but
on Android S devices, this interface was in keymint v2, and so it used
the keymint v2 hash.

In order for that old implementation to be recognized, we need to add the
hash in specifically.

Note: v1 was missed in b/264549860, but this was v2.

Fixes: 271513408
Test: vts_treble_vintf_vendor_test
Change-Id: I58c7c41633000df933261a147edd3477afd09a36