Fix crash issue on test case RadioHidlTest_v1_5#startNetworkScan.
Bad safe_union access is raised by calling the GET method of
eutranBands(), because in default ctor of safe_union RadioAccessSpecifier::Bands
hidl_d is assigned with hidl_discriminator::geranBands, which conflicts
with hidl_discriminator::eutranBands, and leads to crash.

Should use the SET method of eutranBands(&) for assignment purpose.

Bug: 271642958
Change-Id: Ie241e8968eb7f9a1297203be2ab4e0a1bf738dea
Signed-off-by: Jia Jia <jia.jia@zte.com.cn>