Revert "uwb(hal): Implement UCI over serial in the default HAL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c989a768-106a-4ce7-a9f4-fbf3d2b3de7b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0e71160d61568b9a018383cc52e045958331ac00