Revert "uwb(hal): Implement UCI over serial in the default HAL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4796868-8d6b-4eb7-9dc5-246fab510f48). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I873d049461c89cc6196d71c10b9dd772b2d70ebe