Revert "Run VtsHalBluetoothTargetTest with hwasan"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8b07071-5028-49ac-8e80-ed074fc6b3db). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie39144050285fe14b7fb22bcc4473b3db7efd7b5