Revert "Add compatibility matrix for V."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c4e2e8c-66dd-4a92-b27a-e760a06a92b2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icb34125892b171f730a095365ad054c5e2f4a0f1