Bluetooth: Make the H4 parser module standalone
The H4 parsing logic is currently split between the packetizer which
extracts the IDCs and the parser which extracts the header and payload.
That make it complicated to use the parser without the packetizer which
comes with its own socket management logic.

This change reimplements the parser as a separate module,
linked in a separate library.

Test: atest --host bluetooth-vendor-interface-hci-test
Change-Id: Iaafc967c723e4460f46ea2d8ccc0fcb34acc1d6f