Revert "Backport VTS fixes to previous test versions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/10296ac5-a544-473b-8c6a-a6fea1757e0e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4222c3d87a3453280a1c1e070fabbb80dbc150d7