[Proposed Change] Challenge expected in timestamp token in case 2
In the second case out of the two cases of authorization enforcement
described for update(), it seems like the challenge is expected in
the timestamp token.

This update seems to match the reference implementation, except that
if this is correct, we need to pass None for the last argument of
"check_auth_token" call at line 869 in operation.rs.

Test: N/A
Change-Id: I33e1b84bf8218335665b31ca144b3b4ecb342328