Revert "Remove unused hardware/bluetooth.h include"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04345d09-ccc2-4e94-a9f6-12ec646648bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1991cf0af94c334833fb1a6573dcb31a844fa77b