Add capabilities SYS_NICE to cas service
The lsc performance does not match the test case requirement (1Mb/s),
and the MediaCas service doesn't have the privilege to set SYS_NICE
and CI plugin related thread priority is low.
We add the capabilities SYS_NICE to cas service so that we can set
the priority of cas service.

Bug:
Test: manual test
Change-Id: I4d7ee7fae366dc22d986056637e18895c2375dd7