Revert "Only allow cuttlefish to use FCM 9 HALs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d35838f-edaf-44f3-abd5-075e80056aa5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id083120764a8a8aa2cd52cf1d7b5c41897550df1