Revert "Allow fingerprint HAL to be updated via apex in U"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d35838f-edaf-44f3-abd5-075e80056aa5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I383b9a33b5ad761e8adc8d2cc9bd07028e481005