Revert "Only allow cuttlefish to use FCM 9 HALs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a16e4962-8202-4e2a-96c8-51e936d9f579). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05450b145bec5a281464fc02cd1d34d1d106319a