Revert "Allow fingerprint HAL to be updated via apex in U"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a16e4962-8202-4e2a-96c8-51e936d9f579). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7d8b4c675b2a36015e946b4e440439cd450d7880