Let the lazy CAS HAL to override the regular one
Some devices may choose to use the lazy version of CAS HAL than the
regular one, but the default one is added to pretty much all devices in
the commit 4f5534b0d381a0968e81ee20e81b791aebe5493d in the
platform/build project. Adding the lazy version will duplicate the
implementation of CAS HAL.

This is what is done in the 1.2 HIDL implementation.

Bug: 291344726
Test: Only the lazy implementation is installed to devices that require
the lazy version.

Change-Id: I88ee8d2fa4c27720817e65e5c41ffc74d6b2d61f