Introduce Secretkeeper HAL interface
Secretkeeper is a privileged component which seals the secrets of pVM
instances & releases them on successful authentication of the pVM.

The HAL should be backed by a secure TA of higher privilege than a pVM.
This patch also introduces a nonsecure service implementing the
interface - secretkeeper-service.nonsecure, which enables development
without a reference TA implementation. This is the
NonSecureSecretkeeper. Along with that, we add VTS testing the basic
store/read functionality.

Test: atest VtsHalSkTargetTest
Bug: 293429085
Change-Id: I8e650f27d506d378a94bbc8834c68a005fb12253