Document versions of the Android Profile for DICE
Describe the differences between versions of the profile and how the
profile used for each certificate is identified.

The Android Profile for DICE evolves with Android releases. Backwards
compatibility is required to allow updated components to run on
components that don't update, or can't update in the case of ROM.

Bug: 293881506
Test: n/a
Change-Id: I464736bb7f3780941c36e2de452cbba14a9fb330