Revert "AIDL BufferPool implementation (HIDL -> AIDL)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27b1d5e0-ebba-40b1-8ec3-a558e1afb94c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib0a090d833ff59ae29834c9c7fc4b7bc2f2beda1