Revert "AIDL bufferpool implementation (just copy from HIDL impl)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27b1d5e0-ebba-40b1-8ec3-a558e1afb94c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0a5315b57b7f8c78ce2434083b7f1f9d2302505