Revert "bufferpool2: Support mainline s/w codec"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27b1d5e0-ebba-40b1-8ec3-a558e1afb94c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I823b3355fe209abe888d435a0ba938176b553453