Revert "audio: Fix remote submix module device ports handling"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c366f501-9ef8-432e-abdb-c369f54919a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I98c4f0fce7abe4cfc02b665545d9e6e2a830d8e2