Wifi: Fix flaky SupplicantP2pIfaceAidlTest#CancelConnect test
SupplicantP2pIfaceAidlTest#CancelConnect test first issues a
connect request to supplicant with an unknown P2P peer device
address. This triggers a scan in supplicant. And then the test triggers
CancelConnect to cancel the ongoing connection request. The cancelConnect
operation detect a group formation or connection in progress by checking
if there is an ongoing scan request. The cancelConnect will return failure
if the supplicant doesn't schedule the scan immediately.
It is possible that certain supplicant implementations can delay the scan.
On such implementation the CancelConnect returns fail. Since CancelConnect
VTS test cannot really test the end to end P2P functionality, consider
the test as PASS even for a return value of FAILURE_UNKNOWN for
CancelConnect AIDL call.

Bug: 298310537
Test: atest VtsHalWifiSupplicantP2pIfaceTargetTest
Change-Id: Id98bb2761e56c6b9be4173cc26d1b32beb21b941