Revert "Add memtrack HAL APEX"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c77ad52-73d6-417b-974b-7f4e749e32e5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ife56a0f49790065ffc7382a62b47d6bc39c9f95a