Revert "Add contexthub HAL APEX"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/825f64c5-978d-4f36-83cd-f1dfb3387043). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2a410f7d4a4d465253be6b52ac02b64df52fec1