Revert "Add bootcontrol HAL APEX"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/643137bf-8dc7-4c8a-aaee-4c2cc0bfead1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8471a33cb7c504a777a17aa2dca6befb2df4b566